% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IR59}
\alias{IR59}
\title{Total phosphorus concentrations at Illinois River, 2001-2009}
\format{A data frame with three columns and 73 rows.
\describe{
 \item{Date}{The sample dates, should be POSIXct format}
 \item{Flow_cfs}{Streamflow, in cubic feet per second, numeric}
 \item{TP_mgl}{Total phosphorus concentrations, mg per L, numeric}
 ...
}}
\source{
\url{waterdata.usgs.gov}
}
\usage{
IR59
}
\description{
A dataset containing TP (mg/L), flow (cfs), and sample dates from 2001 to
2009. The site is Illinois River south of Siloam Springs, AR; USGS site
07195430 (see waterdata.usgs.gov)
}
\keyword{datasets}

