% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.fitter.R
\name{loess.fitter}
\alias{loess.fitter}
\title{Fit LOESS with K-fold Cross-validation}
\usage{
loess.fitter(x, y, span.vals = seq(0.1, 1, by = 0.05),
  folds = control$folds, degree = control$degree, iternum,
  loss = control$loss)
}
\arguments{
\item{x}{A vector of the independent variable (i.e., log-Q)}

\item{y}{A vector of the dependent variable (i.e., log-C)}

\item{span.vals}{Sequence of possible values of smoothing parameter (span or "f") to evaluate using cross-validation.}

\item{folds}{The number of folds, or random partitions of data, to use in cross-validation.}

\item{degree}{The degree polynomial to use in LOESS. Defaults to linear (degree = 1)}

\item{iternum}{Current iteration index. Used with loess.opt to make sure the C-V is repeatable for evaluating each span value.}

\item{loss}{The loss function to use. Defaults to mean absolute deviation (MAD); can also use mean squared error (MSE), see details.}
}
\description{
Use LOESS, a local regression smoother, to model the CvQ relationship in the data.
}
\seealso{
\code{\link[stats]{loess}}
}

