% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.opt.R
\name{loess.opt}
\alias{loess.opt}
\title{Fit LOESS with iterative K-fold Cross-validation}
\usage{
loess.opt(x, y, control = tafa.control())
}
\arguments{
\item{x}{A vector of the independent variable (i.e., log-Q)}

\item{y}{A vector of the dependent variable (i.e., log-C)}

\item{control}{The control parameters; see \code{\link{tafa.control}}}
}
\description{
Fine-tune LOESS via iterative K-fold CV. Used by tafa to flow-adjust concentrations.
}
\seealso{
\code{\link[stats]{loess}}, \code{\link{tafa}}
}

