% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfxn.R
\name{test}
\alias{test}
\title{A test function}
\usage{
test(zach_angry = TRUE)
}
\arguments{
\item{zach_angry}{Default is- he mad}
}
\description{
This function gives insight to Zach's mental state.
}
\examples{
test()
}
\keyword{FUCK}

